+++ src/js/shapes/iconsShape.js
@@ -1,10 +1,10 @@
 import PropTypes from 'prop-types';
 
 export default PropTypes.shape({
-    moveLeft: PropTypes.node,
-    moveAllLeft: PropTypes.node,
-    moveRight: PropTypes.node,
-    moveAllRight: PropTypes.node,
+    moveToAvailable: PropTypes.node,
+    moveAllToAvailable: PropTypes.node,
+    moveToSelected: PropTypes.node,
+    moveAllToSelected: PropTypes.node,
     moveBottom: PropTypes.node,
     moveDown: PropTypes.node,
     moveUp: PropTypes.node,